requirements.txt
python-telegram-bot
sentence-transformers
sqlite-vec
groq
python-dotenv
numpy